+++ rtic-monotonics/src/lib.rs
@@ -13,3 +13,13 @@ pub mod systick;
 
 #[cfg(feature = "rp2040")]
 pub mod rp2040;
+
+/// This marker is implemented on an interrupt token to enforce that the right tokens
+/// are given to the correct monotonic implementation.
+///
+/// This trait is implemented by this crate and not intended for user implementation.
+///
+/// # Safety
+///
+/// This is only safely implemented by this crate.
+pub unsafe trait InterruptToken<Periperhal> {}