+++ rtic-monotonics/CHANGELOG.md
@@ -9,6 +9,8 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ### Added
 
+- Interrupt tokens for `Systick` and `rp2040` to make sure an interrupt handler exists
+
 ### Changed
 
 ### Fixed